requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
djangorestframework==3.14.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.3.1
pycparser==2.22
PyJWT==2.10.1
pillow
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
djangorestframework-simplejwt
africastalking
django-environ
django-cors-headers
django-redis
django-storages
django-extensions
django-filter
django-debug-toolbar
django-guardian
django-axes
django-redis-cache
django-redis-sessions
django-redis-cluster
whitenoise 
gunicorn 
redis
channels-redis